Refactor: util 경로 수정
