+++ src/components/signupForm/SignupForm.tsx
@@ -7,7 +7,7 @@ TODO - onSubmit 코드 구현할 것.
 import { useForm } from 'react-hook-form'
 import SignInput from '@/components/signInput/SignInput'
 import { SignupFormValueType } from '@/types/formTypes'
-import { emailReg, passwordReg } from '@/utils/validChecker'
+import { emailReg, passwordReg } from '@/utils/regExpressions'
 import styles from './SignupForm.module.scss'
 
 export default function SignupForm() {
